[refactor command] create structure for command pattern
